fix: cast post_id to int from filter_input; remove int type hint from rstore_get_product_meta
filter_input() with FILTER_SANITIZE_NUMBER_INT returns a string, not
an int. post_screen_edit_heading() was passing it directly to
rstore_get_product_meta(int \$post_id) causing a TypeError when editing
any product post. Fix by casting to int at the call site and removing
the strict int type hint from rstore_get_product_meta() so it accepts
string IDs (using absint() internally). Also removes spurious 3rd arg
from Display::wp_enqueue_scripts() call.

test(e2e): add spec 08 to cover admin product edit screen

+++ includes/class-post-type.php
@@ -452,7 +452,7 @@ final class Post_Type {
 
 		}
 
-		$post_id = filter_input( INPUT_GET, 'post', FILTER_SANITIZE_NUMBER_INT );
+		$post_id = (int) filter_input( INPUT_GET, 'post', FILTER_SANITIZE_NUMBER_INT );
 
 		$title = ( $post_id > 0 ) ? rstore_get_product_meta( $post_id, 'title' ) : null;
